Adding output toggle between USB and Bluetooth
The default behavior makes it so that when a USB cable is connected, the
output is also switched to USB even if it's only a charger cable. In
most cases, I am on bluetooth but charging. I'd still like to have the
ability to use a USB cable for keyboard input so I'm using OUT_TOG to be
able to switch.